+++ lib/bull.decorators.ts
@@ -1,4 +1,4 @@
-import { Inject, SetMetadata } from '@nestjs/common';
+import { Inject, SetMetadata, Logger } from '@nestjs/common';
 import { getQueueToken } from './bull.utils';
 import {
   BULL_MODULE_QUEUE,
@@ -15,10 +15,30 @@ import { BullQueueEvent } from './bull.types';
 export const InjectQueue = (name?: string): ParameterDecorator =>
   Inject(getQueueToken(name));
 
-export const Queue = (options?: QueueDecoratorOptions): ClassDecorator =>
+export const Queue = (options?: QueueDecoratorOptions): ClassDecorator => {
+  Logger.warn(
+    `The 'Queue' decorator is deprecated in favor of 'Processor' and will soon be removed.`,
+    'BullModule',
+    false,
+  );
+  return Processor(options);
+};
+
+export const Processor = (options?: QueueDecoratorOptions): ClassDecorator =>
   SetMetadata(BULL_MODULE_QUEUE, options || {});
 
 export const QueueProcess = (
+  options?: QueueProcessDecoratorOptions,
+): MethodDecorator => {
+  Logger.warn(
+    `The 'QueueProcess' decorator is deprecated in favor of 'Process' and will soon be removed.`,
+    'BullModule',
+    false,
+  );
+  return Process(options);
+};
+
+export const Process = (
   options?: QueueProcessDecoratorOptions,
 ): MethodDecorator => SetMetadata(BULL_MODULE_QUEUE_PROCESS, options || {});